refactor: Split operator APIs into different part
